Implement serializer with deduplication
