+++ statham/dsl/elements/base.py
@@ -1,8 +1,6 @@
 # False positive. The cycle exists but is avoided by importing last.
 # pylint: disable=cyclic-import
-import inspect
 from typing import Any, cast, Dict, List, Generic, TypeVar, Union
-from typing_extensions import Literal
 
 from statham.dsl.constants import NotPassed, Maybe
 from statham.dsl.exceptions import ValidationError
@@ -184,52 +182,6 @@ class Element(Generic[T]):
             return value
         return create(value)
 
-    def _serialize(self) -> Dict[str, Any]:
-        """Serialize the DSL Element, without recursion into child elements."""
-        schema = {
-            param.name: getattr(self, param.name, param.default)
-            for param in inspect.signature(
-                type(self).__init__
-            ).parameters.values()
-            if param.kind == param.KEYWORD_ONLY
-            and getattr(self, param.name, param.default) != param.default
-        }
-        if "properties" in schema:
-            schema["required"] = [
-                prop.source
-                for prop in schema["properties"].values()
-                if prop.required
-            ]
-        return schema
-
-    def serialize(self, reference: bool = False) -> Dict[str, Any]:
-        """Serialize the DSL Element.
-
-        :param reference: If True, replace related object schemas
-          with JSON references.
-        """
-        return _serialize_recursive(self._serialize(), reference=reference)
-
-
-def _serialize_recursive(data: Any, reference: bool = False) -> Any:
-    """Recursively serialize DSL elements."""
-    if isinstance(data, Element) and isinstance(data, type) and reference:
-        return f"#/definitions/{data.__name__}"
-    if isinstance(data, _Property):
-        return _serialize_recursive(data.element, reference=reference)
-    if isinstance(data, Element):
-        return data.serialize(reference=reference)
-    if not isinstance(data, (list, dict)):
-        return data
-    if isinstance(data, list):
-        return [
-            _serialize_recursive(item, reference=reference) for item in data
-        ]
-    return {
-        key: _serialize_recursive(value, reference=reference)
-        for key, value in data.items()
-    }
-
 
 UNBOUND_PROPERTY: _Property = _Property(Element(), required=False)
 UNBOUND_PROPERTY.bind_name("<unbound>")
@@ -252,14 +204,6 @@ class Nothing(Element):
     @property
     def validators(self) -> List[Validator]:
         return [NoMatch()]
-
-    def _serialize(self) -> Dict[str, Any]:
-        """Serialize the schema element.
-
-        This uses a long form to keep a uniform interface between
-        elements.
-        """
-        return {"not": True}
 
 
 # Needs to be imported last to prevent cyclic import.
