+++ statham/dsl/elements/string.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List
+from typing import Any, List
 
 from statham.dsl.elements.base import Element
 from statham.dsl.constants import Maybe, NotPassed
@@ -36,6 +36,3 @@ class String(Element[str]):
     @property
     def type_validator(self):
         return InstanceOf(str)
-
-    def _serialize(self) -> Dict[str, Any]:
-        return {**super()._serialize(), "type": "string"}
