+++ statham/serializers/python.py
@@ -1,7 +1,5 @@
-from typing import Any, Dict
-
 from statham.dsl.elements import Element
-from statham.orderer import orderer, get_object_classes
+from statham.orderer import orderer
 
 
 # TODO: Identify which of these to exclude when performing generation.
@@ -34,17 +32,3 @@ def serialize_python(*elements: Element) -> str:
     return _IMPORT_STATEMENTS + "\n\n".join(
         [object_model.python() for object_model in orderer(*elements)]
     )
-
-
-def serialize_json(*elements: Element) -> Dict[str, Any]:
-    """Output JSON Schema dictionary."""
-    primary = elements[0]
-    object_classes = get_object_classes(*elements)
-    return {
-        **primary.serialize(reference=True),
-        "definitions": {
-            object_class.__name__: object_class.serialize(reference=True)
-            for object_class in object_classes
-            if object_class is not primary
-        },
-    }
