+++ tests/dsl/elements/test_array.py
@@ -175,48 +175,3 @@ class TestArrayDefault:
         element = Array(self.Foo, default=[{"other": "foo"}])
         instance = element(NotPassed())
         assert instance == [{"other": "foo"}]
-
-
-@pytest.mark.parametrize(
-    "element,expected",
-    [
-        (Array(Element()), {"type": "array", "items": {}}),
-        (Array(String()), {"type": "array", "items": {"type": "string"}}),
-        (
-            Array(
-                String(),
-                default=["foo", "bar"],
-                minItems=1,
-                maxItems=3,
-                contains=String(),
-                uniqueItems=True,
-            ),
-            {
-                "type": "array",
-                "items": {"type": "string"},
-                "default": ["foo", "bar"],
-                "minItems": 1,
-                "maxItems": 3,
-                "contains": {"type": "string"},
-                "uniqueItems": True,
-            },
-        ),
-        (
-            Array(
-                [String(), Integer()],
-                default=["foo", 1],
-                minItems=2,
-                additionalItems=False,
-            ),
-            {
-                "type": "array",
-                "items": [{"type": "string"}, {"type": "integer"}],
-                "default": ["foo", 1],
-                "minItems": 2,
-                "additionalItems": False,
-            },
-        ),
-    ],
-)
-def test_array_serialize(element: Element, expected: Dict[str, Any]):
-    assert element.serialize() == expected
