+++ tests/dsl/elements/test_boolean.py
@@ -49,14 +49,3 @@ def test_string_default_keyword():
 
 def test_string_type_annotation():
     assert Boolean().annotation == "bool"
-
-
-@pytest.mark.parametrize(
-    "element,expected",
-    [
-        (Boolean(), {"type": "boolean"}),
-        (Boolean(default=True), {"type": "boolean", "default": True}),
-    ],
-)
-def test_boolean_serialize(element: Element, expected: Dict[str, Any]):
-    assert element.serialize() == expected
