+++ tests/dsl/elements/test_composition.py
@@ -236,53 +236,3 @@ def test_default_does_not_affect_sub_schema_attempts():
     with no_raise():
         result = element(NotPassed())
     assert result is NotPassed()
-
-
-def simple_multi_composition_serialize_params():
-    return chain.from_iterable(
-        [
-            [
-                (elem(String()), {keyword: [{"type": "string"}]}),
-                (
-                    elem(String(), default="sample string"),
-                    {keyword: [{"type": "string"}], "default": "sample string"},
-                ),
-                (
-                    elem(String(), Array(String())),
-                    {
-                        keyword: [
-                            {"type": "string"},
-                            {"type": "array", "items": {"type": "string"}},
-                        ]
-                    },
-                ),
-                (
-                    elem(String(), Array(String()), default="sample string"),
-                    {
-                        keyword: [
-                            {"type": "string"},
-                            {"type": "array", "items": {"type": "string"}},
-                        ],
-                        "default": "sample string",
-                    },
-                ),
-            ]
-            for elem, keyword in (
-                (AnyOf, "anyOf"),
-                (AllOf, "allOf"),
-                (OneOf, "oneOf"),
-            )
-        ]
-    )
-
-
-@pytest.mark.parametrize(
-    "element,expected",
-    [
-        (Not(Element()), {"not": {}}),
-        (Not(String()), {"not": {"type": "string"}}),
-        *simple_multi_composition_serialize_params(),
-    ],
-)
-def test_composition_serialize(element, expected):
-    assert element.serialize() == expected
