+++ tests/dsl/elements/test_element.py
@@ -244,60 +244,3 @@ def test_nothing_element_is_falsey():
 
 def test_nothing_element_annotation_is_none():
     assert Nothing().annotation == "None"
-
-
-@pytest.mark.parametrize(
-    "element,expected",
-    [
-        (Element(), {}),
-        (Element(minLength=3), {"minLength": 3}),
-        (
-            Element(
-                default="foo",
-                const="foo",
-                minItems=3,
-                maxItems=5,
-                minimum=3,
-                maximum=5,
-                minLength=3,
-                maxLength=5,
-                required=["value"],
-                properties={"value": _Property(String(), required=True)},
-                additionalProperties=String(),
-                patternProperties={"^foo": String()},
-                minProperties=1,
-                maxProperties=3,
-                items=String(),
-                uniqueItems=True,
-                propertyNames=String(maxLength=3),
-                contains=String(),
-                enum=["foo", "bar"],
-                dependencies={"foo": ["bar"], "qux": Element(minProperties=2)},
-            ),
-            {
-                "default": "foo",
-                "const": "foo",
-                "minItems": 3,
-                "maxItems": 5,
-                "minimum": 3,
-                "maximum": 5,
-                "minLength": 3,
-                "maxLength": 5,
-                "required": ["value"],
-                "properties": {"value": {"type": "string"}},
-                "additionalProperties": {"type": "string"},
-                "patternProperties": {"^foo": {"type": "string"}},
-                "minProperties": 1,
-                "maxProperties": 3,
-                "items": {"type": "string"},
-                "uniqueItems": True,
-                "propertyNames": {"type": "string", "maxLength": 3},
-                "contains": {"type": "string"},
-                "enum": ["foo", "bar"],
-                "dependencies": {"foo": ["bar"], "qux": {"minProperties": 2}},
-            },
-        ),
-    ],
-)
-def test_element_serialize(element: Element, expected: Dict[str, Any]):
-    assert element.serialize() == expected
