+++ tests/dsl/elements/test_object.py
@@ -392,141 +392,3 @@ class TestEnum:
         instance = self.MyObject({"foo": "bar"})
         with no_raise():
             _ = self.MyObject(instance)
-
-
-@pytest.mark.parametrize(
-    "schema,element",
-    [
-        pytest.param(
-            {"type": "object", "title": "EmptyModel"},
-            EmptyModel,
-            id="with-no-keywords",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithOptionalProperty",
-                "properties": {"value": {"type": "string"}},
-            },
-            ObjectWithOptionalProperty,
-            id="with-not-required-property",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithRequiredProperty",
-                "required": ["value"],
-                "properties": {"value": {"type": "string"}},
-            },
-            ObjectWithRequiredProperty,
-            id="with-required-property",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithObjectProperty",
-                "required": ["value"],
-                "properties": {
-                    "value": {
-                        "type": "object",
-                        "title": "ObjectWithRequiredProperty",
-                        "required": ["value"],
-                        "properties": {"value": {"type": "string"}},
-                    }
-                },
-            },
-            ObjectWithObjectProperty,
-            id="with-object-property",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithDefaultProp",
-                "default": {"default": "a string"},
-                "properties": {"value": {"type": "string"}},
-            },
-            ObjectWithDefaultProp,
-            id="with-default-default",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithAdditionalPropElement",
-                "additionalProperties": {"type": "string"},
-            },
-            ObjectWithAdditionalPropElement,
-            id="with-additional-properties-schema",
-        ),
-        pytest.param(
-            {"type": "object", "title": "ObjectWithAdditionalPropTrue"},
-            ObjectWithAdditionalPropTrue,
-            id="with-additional-properties-true",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithAdditionalPropFalse",
-                "additionalProperties": False,
-            },
-            ObjectWithAdditionalPropFalse,
-            id="with-additional-properties-false",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithPatternProps",
-                "patternProperties": {"^foo": {}, "^(?!foo)": False},
-            },
-            ObjectWithPatternProps,
-            id="with-pattern-props",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithSizeValidation",
-                "minProperties": 1,
-                "maxProperties": 2,
-            },
-            ObjectWithSizeValidation,
-            id="with-size-validation",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithPropertyNames",
-                "propertyNames": {"type": "string", "maxLength": 3},
-            },
-            ObjectWithPropertyNames,
-            id="with-property-names",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithConst",
-                "const": {"foo": "bar"},
-            },
-            ObjectWithConst,
-            id="with-const",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithEnum",
-                "enum": [{"foo": "bar"}, {"qux": "mux"}],
-            },
-            ObjectWithEnum,
-            id="with-enum",
-        ),
-        pytest.param(
-            {
-                "type": "object",
-                "title": "ObjectWithDependencies",
-                "dependencies": {"foo": ["bar"], "qux": {"minProperties": 2}},
-            },
-            ObjectWithDependencies,
-            id="with-dependencies",
-        ),
-    ],
-)
-def test_object_serialize(element: ObjectMeta, schema: Dict[str, Any]):
-    assert element.serialize() == schema
