+++ tests/serializers/test_python.py
@@ -9,7 +9,7 @@ from statham.dsl.elements import Element, Object, String
 from statham.dsl.elements import Nothing  # pylint: disable=unused-import
 from statham.dsl.parser import parse
 from statham.dsl.property import Property
-from statham.serializer import _IMPORT_STATEMENTS, serialize_python
+from statham.serializers.python import _IMPORT_STATEMENTS, serialize_python
 
 
 SCHEMA = {
